Menu and Generator Changes
I Changed the way the generator works to keep an index of the last known
empty hexagon so it doesn't search as much and theoretically shorten
load times (probably marginally).

THe menu now draws the map when the menu drops down on top of it.

The menu drops down with delta time to be consistent with frame drops